requirements.txt
Django==2.1.3
django-bootstrap==0.2.4
django-bootstrap4==0.0.7
django-cities-light==3.5.0
django-js-asset==1.1.0
django-mathfilters==0.4.0
django-mptt==0.9.1
django-registration==3.0
opencv-python==3.4.3.18
phonenumbers==8.9.16
Pillow==5.3.0
python-utils==2.3.0
pytz==2018.5
Unidecode==1.0.22
whitenoise==4.1
six==1.11.0
